renamed AppAuthentication to Authentication
